Renaming cifar_transforms to misc_transforms
