+++ fastssl/data/__init__.py
@@ -3,5 +3,5 @@ from .cifar_dataloaders import cifar_ffcv, cifar_classifier_ffcv, cifar_pt
 from .stl10_dataloaders import stl_ffcv, stl10_pt, stl_classifier_ffcv 
 from .imagenet_dataloaders import imagenet_ffcv, imagenet_classifier_ffcv
 from .imagenet_dataloaders_distributed import imagenet_ffcv_dist, imagenet_classifier_ffcv_dist
-from .cifar_transforms import CifarTransform, CifarClassifierTransform, CifarTransformFFCV, STLTransformFFCV
+from .misc_transforms import CifarTransform, CifarClassifierTransform, CifarTransformFFCV, STLTransformFFCV
 from .custom_ffcv_transforms import ColorJitter, RandomGrayscale
